requirements.txt
lxml
requests
feedparser
pluginbase
python-slugify
pymongo
